Rename OptionService -> Optional
Other service implementations do not have the Service suffix.
