+++ tower/src/util.rs
@@ -5,7 +5,7 @@ pub use tower_util::CallAll;
 pub use tower_util::CallAllUnordered;
 pub use tower_util::EitherService;
 pub use tower_util::Oneshot;
-pub use tower_util::OptionService;
+pub use tower_util::Optional;
 pub use tower_util::Ready;
 pub use tower_util::ServiceFn;
 pub use tower_util::UnsyncBoxService;
